Scrape Scripture Chapter Page.ipynb was removed since it was only for testing. It was essentially replaced by Scrape Scripture Verses.ipynb, which is much more elaborate and was used to scrape all verses for all books (e.g. gen, ex, lev, etc) in all volumes (ot, nt, bofm, dc-testament, pgp). scripture_chapter_page.py was extended with new functions ('generate_file_name_root', 'zero_pad', 'scrape_verses_for_book', and 'scrape_verses_for_book_list') used by Scrape Scripture Verses.ipynb for scraping all verses from all books/volumes.
The output files are included in this commit. The file names are structured as follows: <volume index>_<volume id>_<book index>_<book id>.csv. The volume index and book index are both 1-based (i.e. they start at 1 rather than 0). Volume indexes are as follows (where volume ids are in parentheses): 1: Old Testament (ot); 2: New Testament (nt), 3: Book of Mormon (bofm); 4: Doctrine and Covenants (dc-testament); 5: Pearl of Great Price (pgp). As an example, in 3_bofm_13_morm.csv, 3 is the volume index (corresponding to the Book of mormon), 'bofm' is the volume id, 'morm' is the book id corresponding to 'Mormon' in the Book of Mormon and 13 indicates that this is the 13th book in the volume (i.e. in the Book of Mormon). Volume indexes and book indexes are included in the file names to ensure that when listed in alphanumeric order, the file order corresponds to the order in which the books appear in their respective volumes.

In scripture_book_page.py, the function 'parse_h1_id_title1' was updated, because it was not scraping the entire title. This was corrected. Also, the function 'parse_all_text_from_page' was updated to parse title from a different html block if the expected html block was not found. In almost all cases, the expected default block exists, but there are a few examples where it doesn't exist.

Removed ScriptureChapterParser.py, ScriptureChapterParser_old.py, and ScriptureVolumeParser.py, because this is obsolete code that is no longer useful. The new code is cleaner and simpler.

Updated Scrape Scripture Book Information.ipynb to use 'scripture_book_page.py' to get the number of chapters in each book and add these chapter counts to the dataframe of scripture book information. The updated dataframe was exported to a csv file (all_scripture_books.csv).